email  added

+++ family/src/main/java/com/guru/family/entity/Roles.java
@@ -1,9 +1,12 @@
 package com.guru.family.entity;
 
+import javax.persistence.CascadeType;
 import javax.persistence.Entity;
+import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import javax.persistence.ManyToOne;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -17,7 +20,10 @@ public class Roles {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private long id;
+	private Integer id;
 	private String name;
+	
+//	@ManyToOne(cascade = CascadeType.ALL,fetch = FetchType.EAGER)
+//	RegistrationEntity entity;
 
 }
